Started coding up pathfinding for creatures.
It compiles! So obviously everything works, haha. But actually, I haven't really
tested because we don't have any creatures. Also, the current way of doing traceback
is gonna be a bit sketch with multiple creatures, by which I mean it won't work.
Solutions would be (a) do it a better way or (b) only have one creature at a time,
and make them like the boss monster. We'll see.