Extended change: read description
Hey guys, I tried building the thing finally and it was pretty fucked
up in terms of camera perspective. So I attempted to change the
perspective rations and shit to fix it. It made the game a lot larger,
but everything still works. Try this branch out and let me know what
you think. If we don't like it we can always switch it back to being
half sized camera properties, but the build will still be weird.